requirements.txt
flask
pytesseract
pillow
python-dateutil
